# swagger_server/models/cyclo.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Cyclo(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, cyclo_id: str=None, name: str=None, active: bool=True, company_id: str=None):  # noqa: E501
        """Cyclo - a model defined in Swagger

        :param cyclo_id: The cyclo_id of this Cyclo.  # noqa: E501
        :type cyclo_id: str
        :param name: The name of this Cyclo.  # noqa: E501
        :type name: str
        :param active: The active of this Cyclo.  # noqa: E501
        :type active: bool
        :param company_id: The company_id of this Cyclo.  # noqa: E501
        :type company_id: str
        """
        self.swagger_types = {
            'cyclo_id': str,
            'name': str,
            'active': bool,
            'company_id': str
        }

        self.attribute_map = {
            'cyclo_id': 'cyclo_id',
            'name': 'name',
            'active': 'active',
            'company_id': 'company_id'
        }
        self._cyclo_id = cyclo_id
        self._name = name
        self._active = active
        self._company_id = company_id

    @classmethod
    def from_dict(cls, dikt) -> 'Cyclo':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Cyclo of this Cyclo.  # noqa: E501
        :rtype: Cyclo
        """
        return util.deserialize_model(dikt, cls)

    @property
    def cyclo_id(self) -> str:
        """Gets the cyclo_id of this Cyclo.

        Auto-generated primary key field  # noqa: E501

        :return: The cyclo_id of this Cyclo.
        :rtype: str
        """
        return self._cyclo_id

    @cyclo_id.setter
    def cyclo_id(self, cyclo_id: str):
        """Sets the cyclo_id of this Cyclo.

        Auto-generated primary key field  # noqa: E501

        :param cyclo_id: The cyclo_id of this Cyclo.
        :type cyclo_id: str
        """
        if cyclo_id is None:
            raise ValueError("Invalid value for `cyclo_id`, must not be `None`")  # noqa: E501

        self._cyclo_id = cyclo_id

    @property
    def name(self) -> str:
        """Gets the name of this Cyclo.


        :return: The name of this Cyclo.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this Cyclo.


        :param name: The name of this Cyclo.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def active(self) -> bool:
        """Gets the active of this Cyclo.


        :return: The active of this Cyclo.
        :rtype: bool
        """
        return self._active

    @active.setter
    def active(self, active: bool):
        """Sets the active of this Cyclo.


        :param active: The active of this Cyclo.
        :type active: bool
        """

        self._active = active

    @property
    def company_id(self) -> str:
        """Gets the company_id of this Cyclo.

        This property is a reference to a Company  # noqa: E501

        :return: The company_id of this Cyclo.
        :rtype: str
        """
        return self._company_id

    @company_id.setter
    def company_id(self, company_id: str):
        """Sets the company_id of this Cyclo.

        This property is a reference to a Company  # noqa: E501

        :param company_id: The company_id of this Cyclo.
        :type company_id: str
        """
        if company_id is None:
            raise ValueError("Invalid value for `company_id`, must not be `None`")  # noqa: E501

        self._company_id = company_id